requirements.txt
torch
torchvision
numpy<2.0.0
opencv-python-headless
pandas
matplotlib
flask
scikit-learn
pillow
scikit-image
gunicorn
python-dotenv
fpdf2
